feat: shorten insight share links
