+++ ui/homepage/app/insight/[shareId]/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '../InsightWorkbench'
